pyproject.toml
[tool.poetry]
name = "deep-learning-studies"
version = "0.1.0"
description = ""
authors = ["Eduardo Zampiere <carloszampiere@gmail.com>"]
readme = "README.md"
packages = [{include = "deep_learning_studies"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11.0"
pandas = "^2.0.0"
scikit-learn = "^1.2.2"
tensorflow-macos = "^2.12.0"
matplotlib = "^3.7.1"
scikeras = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
